Add a driver crate
